requirements.txt
Django~=5.0.6
psycopg==3.1.19
Pillow>=10.3.0,<10.4.0
python-decouple==3.8
djangorestframework==3.15.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
pytest-django==4.5.2
drf-spectacular>=0.27.2,<0.28.0
fontawesomefree==5.15.4
drf-nested-routers==0.94.1